Move serialization tests to the new serialization crate
